masValidaciones_primer-preEntrega

+++ src/routes/products.router.js
@@ -3,6 +3,7 @@ const express = require("express");
 const router = express.Router();
 
 const productTest = new ProductManager("./src/models/array-product.json");
+//Actividades desafío3
 //Establezco la ruta products & limit
 router.get("/", async (req, res) => {
   try {
@@ -37,7 +38,8 @@ router.get("/:id", async (req, res) => {
   }
 });
 
-//Agrego producto
+//Actividades 1° pre-entrega
+//Ruta para agregar productos
 router.post("/", async (req, res) => {
   try {
     //Capturo datos que vienen del body
@@ -56,25 +58,43 @@ router.post("/", async (req, res) => {
   }
 });
 
-//Actualizo producto por ID
+//Ruta para actualizar producto según ID especificado
 router.put("/:id", async (req, res) => {
   try {
     const productId = parseInt(req.params.id);
     const putProductBody = req.body;
 
+    //Validación para verificar si existe  el producto correspondiente al ID especificado
+    const productExists = await productTest.productExists(productId);
+    if (!productExists) {
+      return res.status(404).send({
+        message:
+          "El producto no fue encontrado. Verificar el identificador ingresado",
+      });
+    }
+
     await productTest.updateProduct(productId, putProductBody);
 
     res.status(201).send({ message: "Producto actualizado correctamente" });
   } catch (error) {
-    res.status(500).send({ message: "Error al actualizar el producto" });
+    res.status(500).send({ message: "Error al actualizar el producto." });
   }
 });
 
-//Elimino producto por ID
+//Ruta para eliminar producto según ID especificado
 router.delete("/:id", async (req, res) => {
   try {
     const productId = parseInt(req.params.id);
     const putProductBody = req.body;
+
+    //Validación para verificar si existe  el producto correspondiente al ID especificado
+    const productExists = await productTest.productExists(productId);
+    if (!productExists) {
+      return res.status(404).send({
+        message:
+          "El producto no fue encontrado. Verificar el identificador ingresado",
+      });
+    }
 
     await productTest.deleteProduct(productId, putProductBody);
